Gestion des appels depuis un chemin autre que le cwd
git-svn-id: file:///home/ahjucaf/svn/ahjucaf/trunk@226 59802511-991d-49c3-8a85-851a450bd420

+++ juricaf2solr/bin/updateCouchdbView.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
+
+PREDIR=.
+if echo $0 | grep '/' > /dev/null; then PREDIR=$(echo $0 | sed 's/\/[^\/]*$//'); fi
+if ! test -e $PREDIR/../conf/juricaf.conf; then
+    echo $PREDIR/../conf/juricaf.conf does not exist
+    exit 1;
+fi
+. $PREDIR/../conf/juricaf.conf
+
 LOCK=$0.lock 
 if test -e $LOCK ; then
         exit 1;
 fi
 
 touch $LOCK
-wget -q -O /dev/null http://localhost:5984/ahjucaf/_design/stats/_view/pays_juridiction_date
+wget -q -O /dev/null $COUCHDBURL/_design/stats/_view/pays_juridiction_date
 rm $LOCK
